# openapi.yaml
openapi: 3.0.3
info:
  title: 焚信 BurnMsg API
  version: 2.0.0
  description: |-
    焚信（BurnMsg）是一款企业内部加密即时通讯（IM）应用，支持云聊天与秘密聊天、阅后即焚消息。本文件为其后端（NestJS + TypeORM + MySQL）的标准 OpenAPI 3.0.3 接口文档。

    【账号体系】
    - 不开放自助注册，所有账号由管理后台统一开通（POST /accounts，仅 admin）。
    - 用户名即公司内部手机号（1 开头 11 位数字）。
    - 新开通 / 被重置密码的账号 force_change_pwd=true，前端应强制跳转修改密码，修改成功后该标记置为 false。

    【认证方式】
    - 受保护接口在请求头携带：Authorization: Bearer <access_token>。
    - 双 token 机制：access_token（JWT_SECRET 签名，默认有效期 7200 秒）+ refresh_token（独立的 JWT_REFRESH_SECRET 签名，默认有效期 2592000 秒 / 30 天，payload 带 type=refresh 标记防混用）。
    - access_token 过期后用 refresh_token 调 POST /auth/refresh-token 换取新 token 对。
    - 即时吊销：JWT 守卫除验签外，每个受保护请求都会实时查库校验「账号仍为 active」且「签发 token 的设备记录仍存在」；账号被停用或设备被下线后，已签发的 token 立即失效（全端即时强制下线）。

    【统一响应格式】
    - 成功：{ "code": 0, "message": "ok 或各接口自定义提示语", "data": <业务数据，无内容时为 null> }
    - 失败：{ "code": <HTTP 状态码>, "message": "<错误描述；参数校验失败时为多条错误以 '; ' 拼接>" }

    【通用约定】
    - 分页参数：page（默认 1）、pageSize（默认 20，最大 200；通讯录列表默认 50）。
    - 请求中的时间字段使用 ISO8601 字符串；响应中的时间为服务端 datetime。
    - 所有业务 ID 均为 UUID v4 字符串。
    - 角色：user / admin。标注「仅管理员」的接口仅 admin 可调用，普通用户返回 403。
    - 强制改密拦截：force_change_pwd=true 的账号（新开通 / 被重置密码后），除 POST /auth/change-password 与 GET /auth/profile 外，调用其他任何受保护接口一律返回 403（FORCE_CHANGE_PASSWORD），前端应引导其先完成改密再进入主界面。
    - 软删除账号：被 DELETE /accounts/{id} 软删的账号立即吊销全部 token 并强制下线，其历史消息保留、前端统一展示「已注销用户」；账号列表默认不含已软删账号，加 show_deleted=1 可查看。
servers:
  - url: http://localhost:9091/api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /groups:
    post:
      tags: [群组]
      summary: 创建群组或频道，创建者为 owner
      description: |-
        【用途】创建群组（或频道）会话。

        【调用时机】用户在 App 选择「发起群聊/创建频道」并选定初始成员后提交时。

        【业务规则与注意事项】
        - 创建者自动成为群主（role=owner），同时计入成员；member_ids 中即使包含创建者也会被自动去重。
        - member_ids 最多 500 个，每个必须是合法 UUID v4；新加成员角色均为 member。
        - is_channel=true 时创建频道：会话 is_channel 标记为 true（源码中会话 type 恒为 group，频道通过 is_channel 字段区分）。
        - 创建成功后 member_count = 1（创建者）+ 实际初始成员数。
        - 源码未校验 member_ids 对应的用户是否存在/启用，调用方应传入从通讯录获取的有效用户 ID。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 200
                  description: 群名称，最长 200 字符
                  example: 产品评审群
                description:
                  type: string
                  description: 群简介（可选）
                  example: 每周三产品评审同步
                avatar_url:
                  type: string
                  maxLength: 500
                  description: 群头像 URL（可选，可先通过 POST /upload 上传获得），最长 500 字符
                  example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
                member_ids:
                  type: array
                  maxItems: 500
                  items: { type: string, format: uuid }
                  description: 初始成员用户 ID 数组（可选，最多 500 个；创建者无需包含在内，重复 ID 自动去重）
                  example: ['7c9e6679-7425-40de-944b-e07fc1f90ae7']
                is_channel:
                  type: boolean
                  description: 是否创建为频道（默认 false 为普通群组）
                  example: false
      responses:
        '200':
          description: 创建成功（message 为「群组创建成功」），data 为会话 Conversation（type=group）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 群组创建成功 }
                  data:
                    $ref: '#/components/schemas/Conversation'
        '400':
          description: 参数校验失败（name 缺失或超长、member_ids 超 500 个或含非法 UUID 等）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/{id}:
    put:
      tags: [群组]
      summary: 更新群资料（均为可选字段）
      description: |-
        【用途】更新群名称、简介、头像，三个字段均为可选，仅提交的字段会被更新。

        【调用时机】群主/管理员在群设置页修改资料保存时。

        【业务规则与注意事项】
        - 仅群主（owner）或群管理员（admin）可调用，普通成员返回 403（Only group owner or admin can update group info）。
        - 响应为更新后的完整 Conversation。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
                  description: 群名称，最长 200 字符
                  example: 产品评审群（新版）
                description:
                  type: string
                  description: 群简介
                  example: 每周三 14:00 产品评审
                avatar_url:
                  type: string
                  maxLength: 500
                  description: 群头像 URL，最长 500 字符
                  example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
      responses:
        '200':
          description: 更新成功（message 为「群组信息更新成功」），data 为更新后的 Conversation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 群组信息更新成功 }
                  data:
                    $ref: '#/components/schemas/Conversation'
        '400':
          description: 参数校验失败（如字段超长）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 权限不足，仅群主/群管理员可操作（Only group owner or admin can update group info）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [群组]
      summary: 群主解散群（解散即焚）
      description: |-
        【用途】群主解散自己创建的群。与管理员强制解散（DELETE /groups/admin/{id}）语义不同：本接口为「解散即焚」。

        【调用时机】群主在群设置页确认解散时。

        【业务规则与注意事项】
        - 仅该群 owner 可调用，其他成员/群管理员返回 403；已解散的群返回 400。
        - 解散即焚：解散瞬间全群消息被置为立即到期（destroy_at=解散时刻），消息列表对成员立即不可见，约 1 分钟内由服务端焚毁调度器物理清除（含回执与附件文件）。
        - 群记录（Conversation）保留并打 dissolved_at + dissolved_by 标记，成员不可再发消息；群内成员 WS 实时收到 conversation:updated（reason=dissolved）。
        - 与「阅后即焚」共用同一条物理清除链路，清除动作不可恢复，前端交互上应二次确认。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      responses:
        '200':
          description: 解散成功（message 为「群组已解散」），data 为 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 群组已解散 }
                  data: { nullable: true, example: null }
        '400':
          description: 群已解散 / 目标不是群会话
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 仅群 owner 可解散该群
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 群不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/admin/all:
    get:
      tags: [群组]
      summary: 全量群组管理列表（仅 admin）
      description: |-
        【用途】管理员分页查看全平台所有群组，含群主信息、成员数与解散状态。

        【调用时机】管理后台「群组管理」列表页加载与筛选时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用。
        - 默认不含已解散群，传 include_dissolved=true 可一并查看（dissolved_at 非空即为已解散）。
        - 响应分页结构为 { data, total }，data 元素在 Conversation 基础上补充 owner_name（群主显示名）与 member_count。
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 页码，默认 1
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
          description: 每页条数，默认 20
        - name: keyword
          in: query
          schema: { type: string }
          description: 群名称模糊搜索
          example: 产品
        - name: include_dissolved
          in: query
          schema: { type: string, enum: ['true'] }
          description: 传 true 时结果包含已解散群；默认不含
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        description: 群组列表（Conversation + owner_name + member_count + dissolved_at/dissolved_by）
                        items:
                          $ref: '#/components/schemas/Conversation'
                      total:
                        type: integer
                        description: 满足条件的群组总数
                        example: 42
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/admin/{id}:
    delete:
      tags: [群组]
      summary: 管理员强制解散群（留痕，仅 admin）
      description: |-
        【用途】管理员强制解散任意群组，用于违规群处置。与群主解散（DELETE /groups/{id}）语义不同：本接口「留痕」——消息记录保留供审计，不做物理焚毁。

        【调用时机】管理后台「群组管理」对违规群执行强制解散时。

        【业务规则与注意事项】
        - 仅 admin 角色可调用；已解散的群重复操作返回 400。
        - 解散后群打 dissolved_at + dissolved_by（操作管理员）标记，成员不可再发消息，消息与回执完整保留，可通过审计/管理后台追溯。
        - 群内成员 WS 实时收到 conversation:updated（reason=dissolved），前端应将群置灰并禁止输入。
        - 审计日志留痕（action=dissolve_group，含操作者与 IP）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      responses:
        '200':
          description: 解散成功（message 为「群组已解散」），data 为 null
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 群组已解散 }
                  data: { nullable: true, example: null }
        '400':
          description: 群已解散 / 目标不是群会话
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非 admin 角色（Admin access required）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 群不存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/{id}/members:
    get:
      tags: [群组]
      summary: 群成员列表（含成员用户信息）
      description: |-
        【用途】获取群成员列表，每条成员记录附带该成员的姓名与头像。

        【调用时机】打开群聊的成员列表/群设置页时。

        【业务规则与注意事项】
        - 仅本群成员可查看，非成员返回 403（修复了旧版任意用户可查漏洞）。
        - 结果按成员角色 role 字段升序排列；每条在 ConversationMember 基础上附带 user_display_name、user_avatar_url 便于直接展示。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      responses:
        '200':
          description: 成功，data 为群成员数组
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: ok }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupMember'
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非群成员（Access denied to this conversation）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    post:
      tags: [群组]
      summary: 批量添加群成员
      description: |-
        【用途】向群内批量添加成员。

        【调用时机】群主/管理员在群设置页「添加成员」选人后提交时。

        【业务规则与注意事项】
        - 仅群主（owner）或群管理员（admin）可调用，普通成员返回 403（Only group owner or admin can add members）。
        - member_ids 必填，最多 500 个，每个必须是合法 UUID v4；新成员角色均为 member。
        - 已在群内的用户 ID 自动跳过（幂等），全部已存在时不做任何变更；添加成功后同步更新会话 member_count。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [member_ids]
              properties:
                member_ids:
                  type: array
                  maxItems: 500
                  items: { type: string, format: uuid }
                  description: 待添加的用户 ID 数组（最多 500 个）
                  example: ['7c9e6679-7425-40de-944b-e07fc1f90ae7']
      responses:
        '200':
          description: 添加成功（message 为「成员添加成功」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 成员添加成功 }
                  data: { type: object, nullable: true, example: null }
        '400':
          description: member_ids 缺失、超 500 个或含非法 UUID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 权限不足，仅群主/群管理员可操作（Only group owner or admin can add members）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/{id}/members/{userId}:
    delete:
      tags: [群组]
      summary: 移除群成员
      description: |-
        【用途】把指定用户移出群聊。

        【调用时机】群主/管理员在成员列表中执行「移出群聊」时。

        【业务规则与注意事项】
        - 仅群主（owner）或群管理员（admin）可调用（Only group owner or admin can remove members）；源码未对「本人退群」做单独豁免，普通成员调用同样返回 403。
        - 群主不可被移除，尝试移除群主返回 403（Cannot remove the group owner）。
        - 移除成功后同步更新会话 member_count。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 被移除用户 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      responses:
        '200':
          description: 移除成功（message 为「成员已移除」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 成员已移除 }
                  data: { type: object, nullable: true, example: null }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 权限不足，仅群主/群管理员可操作；或尝试移除群主（Cannot remove the group owner）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /groups/{id}/members/{userId}/role:
    put:
      tags: [群组]
      summary: 设置群成员角色
      description: |-
        【用途】设置指定群成员的角色（设为管理员或降为普通成员）。

        【调用时机】群主在成员列表中「设为管理员/取消管理员」时。

        【业务规则与注意事项】
        - 仅群主（owner）可调用，群管理员也无权设置角色（Only group owner can set admin role）。
        - role 仅允许 admin / member 两种取值（群主身份不可通过本接口转让）。
        - 目标用户必须在群内，否则返回 404（目标用户不在群内）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 群会话 ID（UUID）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: 目标用户 ID（UUID）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
                  description: 目标角色：admin 设为管理员 / member 降为普通成员
                  example: admin
      responses:
        '200':
          description: 设置成功（message 为「角色设置成功」，data 为 null）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 0 }
                  message: { type: string, example: 角色设置成功 }
                  data: { type: object, nullable: true, example: null }
        '400':
          description: role 缺失或取值非法（仅允许 admin/member）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: 未携带或无效/过期 token；账号已停用或设备已下线
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 非群主（Only group owner can set admin role）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: 目标用户不在群内
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        JWT access token。登录（POST /auth/login）或刷新（POST /auth/refresh-token）后获得，放入请求头：
        Authorization: Bearer <access_token>
        默认有效期 7200 秒；每个受保护请求都会实时校验账号状态与设备存在性，账号停用或设备下线后 token 立即失效。
  schemas:
    Error:
      type: object
      description: 统一错误响应（code 为 HTTP 状态码；参数校验失败时 message 为多条错误以 '; ' 拼接）
      properties:
        code:
          type: integer
          description: HTTP 状态码（400/401/403/404/409/413/429/500）
          example: 400
        message:
          type: string
          description: 错误描述
          example: 手机号格式不正确
      required: [code, message]
    SafeUser:
      type: object
      description: 对外用户信息（字段白名单，任何接口不会返回 password_hash 等敏感字段）
      properties:
        id:
          type: string
          format: uuid
          description: 用户 ID（UUID v4）
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        phone:
          type: string
          description: 手机号（即用户名），1 开头 11 位数字
          example: '13800138000'
        display_name:
          type: string
          description: 显示名/姓名
          example: 张三
        avatar_url:
          type: string
          nullable: true
          description: 头像 URL，未设置为 null
          example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
        signature:
          type: string
          nullable: true
          description: 个性签名，未设置为 null
          example: 保持热爱，奔赴山海
        department:
          type: string
          nullable: true
          description: 部门，未设置为 null
          example: 工程部
        role:
          type: string
          enum: [user, admin]
          description: 角色：user 普通用户 / admin 管理员（可调用账号管理接口）
          example: user
        status:
          type: string
          enum: [active, disabled]
          description: 账号状态：active 正常 / disabled 已停用（停用后 token 立即失效且无法登录）
          example: active
        force_change_pwd:
          type: boolean
          description: 是否需强制改密；true 时前端应强制跳转修改密码（新开通或被重置密码的账号为 true）
          example: false
        created_at:
          type: string
          format: date-time
          description: 创建时间（服务端 datetime）
          example: '2026-08-01T02:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: 更新时间（服务端 datetime）
          example: '2026-08-20T09:30:00.000Z'
...
    Conversation:
      type: object
      description: 会话对象（私聊/群/频道；私聊在会话列表中附带 other_user）
      properties:
        id:
          type: string
          format: uuid
          description: 会话 ID（UUID v4）
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        type:
          type: string
          enum: [private, group, channel]
          description: 会话类型：private 私聊 / group 群组 / channel 频道（源码建群时 type 恒为 group，频道以 is_channel 字段区分）
          example: group
        name:
          type: string
          nullable: true
          description: 会话名称（私聊为 null，前端用 other_user.display_name 展示）
          example: 产品评审群
        description:
          type: string
          nullable: true
          description: 群简介
          example: 每周三产品评审同步
        avatar_url:
          type: string
          nullable: true
          description: 会话头像 URL
          example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
        owner_id:
          type: string
          format: uuid
          nullable: true
          description: 群主用户 ID（群会话），私聊为 null
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        is_channel:
          type: boolean
          description: 是否频道
          example: false
        member_count:
          type: integer
          description: 成员数
          example: 12
        last_message_at:
          type: string
          format: date-time
          nullable: true
          description: 最后一条消息时间（会话列表按此倒序，无消息为 null 排最后）
          example: '2026-08-26T08:00:00.000Z'
        dissolved_at:
          type: string
          format: date-time
          nullable: true
          description: 解散时间；非空即已解散（成员不可再发消息），未解散为 null
          example: null
        dissolved_by:
          type: string
          format: uuid
          nullable: true
          description: 解散操作人用户 ID（群主解散=群主，管理员强制解散=管理员），未解散为 null
          example: null
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: '2026-08-01T02:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: '2026-08-26T08:00:00.000Z'
        other_user:
          allOf:
            - $ref: '#/components/schemas/SafeUser'
          nullable: true
          description: 私聊对方用户信息（仅 GET /conversations 会话列表中 type=private 时附带）
    ConversationMember:
      type: object
      description: 会话成员对象
      properties:
        id:
          type: string
          format: uuid
          description: 成员记录主键
          example: 2a3b4c5d-6e7f-48a9-b0c1-d2e3f4a5b6c7
        conversation_id:
          type: string
          format: uuid
          description: 所属会话 ID
          example: 1b2c3d4e-5f6a-47b8-89c0-d1e2f3a4b5c6
        user_id:
          type: string
          format: uuid
          description: 成员用户 ID
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        role:
          type: string
          enum: [owner, admin, member]
          description: 群内角色：owner 群主（创建者，可设置管理员）/ admin 群管理员（可改群资料、增删成员）/ member 普通成员
          example: member
        muted:
          type: boolean
          description: 是否免打扰
          example: false
        last_read_message_id:
          type: string
          format: uuid
          nullable: true
          description: 最近已读到的消息 ID（标记已读时更新）
          example: 5e6f7a8b-9c0d-41e2-83f4-a5b6c7d8e9f0
        joined_at:
          type: string
          format: date-time
          description: 加入时间
          example: '2026-08-01T02:00:00.000Z'
    GroupMember:
      type: object
      description: 群成员列表项（ConversationMember + 成员用户的姓名与头像）
      allOf:
        - $ref: '#/components/schemas/ConversationMember'
        - type: object
          properties:
            user_display_name:
              type: string
              description: 成员显示名（用户不存在时为 Unknown）
              example: 张三
            user_avatar_url:
              type: string
              nullable: true
              description: 成员头像 URL
              example: /uploads/9b1d2c3e-4f5a-4b6c-8d7e-9f0a1b2c3d4e.jpg
...